+++ src/gfn/gym/helpers/bayesian_structure/prior.py
