+++ src/gfn/gym/helpers/bayesian_structure/scores.py
@@ -7,8 +7,8 @@ import torch
 from scipy.special import gammaln
 from torch_geometric.utils import to_dense_adj
 
-from src.gfn.gym.helpers.prior import BasePrior
-from src.gfn.states import GraphStates
+from gfn.gym.helpers.bayesian_structure.prior import BasePrior
+from gfn.states import GraphStates
 
 
 def logdet(matrix: torch.Tensor) -> torch.Tensor:
